social_media_api/requirements.txt
# requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
Pillow==10.1.0
django-filter==23.5